Add 3 layer controller, services, repository, of search property

+++ src/main/java/com/javaweb/repository/ListingRepository.java
@@ -1,12 +1,15 @@
 package com.javaweb.repository;
 
+
 import com.javaweb.repository.entity.ListingEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
 import java.util.Optional;
 
+
 @Repository
-public interface ListingRepository extends JpaRepository<ListingEntity, Long> {
-    Optional<ListingEntity> findByPropertyId(Long propertyId);
+public interface ListingRepository extends JpaRepository<ListingEntity, Integer> {
+    Optional<ListingEntity> findByPropertyId(Integer propertyId);
 }
